+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # PayPal Android SDK Release Notes
 
+## unreleased
+* `PayPalNativePayments`:
+  *  Bump `PayPal Native Checkout` to `0.8.8` and add `return_url`
+
 ## 0.0.7 (2023-01-25)
 * Rename `PayPalDataCollector` to `FraudProtection`
 * Rename `PayPalNativeCheckout` to `PayPalNativePayments`